re-packed SIGIL data sets archive, including some new datasets
git-svn-id: svn://svn.r-forge.r-project.org/svnroot/sigil@76 edb9625f-4e0d-4859-8d74-9fd3b1da38cb

+++ datasets/prepare_data.R
@@ -114,6 +114,11 @@ save(BNCqueries, file="rda/BNCqueries.rda", compress="xz")
 
 
 ## make ZIP archive containing all data files
-zip.name <- "sigil_datasets.zip"
-if (file.exists(zip.name)) stopifnot(file.remove(zip.name))
-system2("zip", c("-r", zip.name, "tbl/*"))
+if (FALSE) {
+  ## should rather pack selected files manually
+  zip.name <- "sigil_datasets.zip"
+  if (file.exists(zip.name)) stopifnot(file.remove(zip.name))
+  setwd("tbl")
+  system2("zip", c("-r", paste0("../", zip.name), "*"))
+  setwd("..")
+}
